Massive updates
Framework overhaul, updating and fixing weird php issues to run on local
dev server.

should now effectively autoload sentry from vendors/autoloader.php

Marco -- if you want to implement the autoloader in a different way,
make the changes and submit. I was having issues including the sentry
class using the files in the src/ directory, since there was no
autoloader. I reverted to this for the meantime since it seems to work.