Fixes to see controller gifs
Implemented controller gifs to be pulled automatically via AJAX. Data
is stored in api/frameData.php and is automatically pulled to couple a
gif on a technique page if the technique is specified in
techs/controller.php in the array $hasFrameData. As we add frameData
for more pages, simply add the json arrays in api/frameData.php and
include the technique in $hasFrameData.