Added sentry group checking to admin page
This is set by visiting techs/god_mode.php on an account that has been
flagged (in techs/mods.php) as a moderator. If this user is pre
approved, then hitting the god mode url will officially flag their
account as an admin - this will display the moderate link in the
navbar.